requirementsbizsystems.txt
Babel==2.1.1
Django==1.9
Django-Accounts==0.1
Fabric==1.10.1
Markdown==2.6.5
Pillow==2.7.0
Shapely==1.5.13
amqp==1.4.8
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.22
celery==3.1.19
decorator==4.0.6
django-allauth==0.24.1
django-audit==1.0b3
django-bootstrap-calendar==0.1.0
django-bootstrap-toolkit==2.15.0
django-bootstrap3==6.2.2
django-celery==3.1.17
django-chartit==0.1
django-crispy-forms==1.5.2
django-email-notification==1.3
django-geojson==2.8.1
django-grappelli==2.7.2
django-leaflet==0.17.1
django-markdown==0.8.4
django-phonenumber-field==1.0.0
django-registration==2.0.3
django-suit==0.2.15
djangorestframework==3.3.2
easy-thumbnails==2.3
ecdsa==0.13
icalendar==3.9.2.dev0
ipython==4.0.1
ipython-genutils==0.1.0
jsmin==2.2.0
kombu==3.0.32
oauthlib==1.0.3
paramiko==1.16.0
path.py==8.1.2
pexpect==4.0.1
phonenumbers==7.2.2
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
pycrypto==2.6.1
pymongo==3.2
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.7
reportlab==3.2.0
requests==2.9.1
requests-oauthlib==0.6.0
simplegeneric==0.8.1
simplejson==3.8.1
six==1.10.0
traitlets==4.0.0
vectorformats==0.1
wsgiref==0.1.2
